Create README.md
The project aims to build a loyalty points management system on Salesforce platform. Salesforce is a customer relation management platform. A CRM in Salesforce works by managing the standard objects, and maintaining the relationships between them, and the standard in-built functionalities. 
The custom fields will be the column of detail we want to keep of a purchase made by a customer inside an object. Next will be defining relationships between the fields, inserting formulas where required. The object is the shortcuts inside the application which lets you access the records and fields inside it. The record is the individual customer entry in the table.